+++ src/agrum/learning/aprioris/aprioriSmoothing.h
@@ -39,7 +39,7 @@ namespace gum {
      * @ingroup learning_group
      */
     template <typename IdSetAlloc = std::allocator<unsigned int>,
-              typename CountAlloc = std::allocator<float>>
+              typename CountAlloc = std::allocator<double>>
     class AprioriSmoothing : public Apriori<IdSetAlloc, CountAlloc> {
       public:
       using type = AprioriSmoothingType;
