+++ src/agrum/learning/aprioris/aprioriSmoothing.tcc
@@ -72,7 +72,7 @@ namespace gum {
         const unsigned int size = this->_target_nodesets->size();
         for ( unsigned int i = 0; i < size; ++i ) {
           if ( this->_target_nodesets->operator[]( i ) != nullptr ) {
-            std::vector<float, CountAlloc>& apriori =
+            std::vector<double, CountAlloc>& apriori =
                 this->_apriori_counts[this->_target_nodesets->operator[]( i )
                                           ->second];
             for ( auto& count : apriori ) {
@@ -81,11 +81,11 @@ namespace gum {
           }
 
           if ( this->_conditioning_nodesets->operator[]( i ) != nullptr ) {
-            const float weight =
+            const double weight =
                 this->_weight *
                 ( *( this->_modalities ) )[( *( this->_target_nodesets ) )[i]
                                                ->first.back()];
-            std::vector<float, CountAlloc>& apriori =
+            std::vector<double, CountAlloc>& apriori =
                 this->_apriori_counts[this->_conditioning_nodesets->operator[](
                                                                       i )
                                           ->second];
