+++ src/agrum/learning/database/DBCellTranslators/cellTranslatorNumber.cpp
@@ -35,7 +35,7 @@ namespace gum {
   namespace learning {
 
     /// default constructor
-    CellTranslatorNumber::CellTranslatorNumber( Sequence<float> values,
+    CellTranslatorNumber::CellTranslatorNumber( Sequence<double> values,
                                                 bool check_database )
         : __check_database( check_database ) {
 
@@ -53,7 +53,7 @@ namespace gum {
       } else {
         // if we specified values, store them
         if ( !values.empty() ) {
-          __user_values = new Sequence<float>( std::move( values ) );
+          __user_values = new Sequence<double>( std::move( values ) );
         }
       }
     }
@@ -63,12 +63,12 @@ namespace gum {
       if ( __user_values != nullptr ) {
         // we probably need to reorder the sequence to be compatible with the
         // order of the values specified by the user
-        std::vector<float> no_user_vals;
-        std::vector<std::pair<float, unsigned int>> user_vals;
+        std::vector<double> no_user_vals;
+        std::vector<std::pair<double, unsigned int>> user_vals;
         for ( unsigned int i = 0; i < __max_value; ++i ) {
-          const float val = __values.first( i );
+          const double val = __values.first( i );
           if ( __user_values->exists( val ) ) {
-            user_vals.push_back( std::pair<float, unsigned int>(
+            user_vals.push_back( std::pair<double, unsigned int>(
                 val, __user_values->pos( val ) ) );
           } else {
             no_user_vals.push_back( val );
@@ -78,8 +78,8 @@ namespace gum {
         // reorder user_vals in increasing order of the second argument
         std::sort( user_vals.begin(),
                    user_vals.end(),
-                   []( const std::pair<float, unsigned int>& elt1,
-                       const std::pair<float, unsigned int>& elt2 ) -> bool {
+                   []( const std::pair<double, unsigned int>& elt1,
+                       const std::pair<double, unsigned int>& elt2 ) -> bool {
                      return elt1.second < elt2.second;
                    } );
 
@@ -103,7 +103,7 @@ namespace gum {
 
     /// specify the set of possible values (to do before creating the row
     /// filter)
-    void CellTranslatorNumber::setUserValues( const Sequence<float>& values,
+    void CellTranslatorNumber::setUserValues( const Sequence<double>& values,
                                               bool check_database ) {
       // clear all current data
       __values.clear();
@@ -130,7 +130,7 @@ namespace gum {
       } else {
         // if we specified values, store them
         if ( !values.empty() ) {
-          __user_values = new Sequence<float>( values );
+          __user_values = new Sequence<double>( values );
         }
       }
     }
